+++ docker/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the docker containers will be documented in this file.
 
+### 2025-11-03
+- Set statsId and readonly for scan policies.
+
 ### 2025-10-31
 - Added config options for recording which packaged scan is being run.